chinh sua chuc nang tinh hinh su dung lao dong

+++ app/Http/Controllers/UserController.php
@@ -23,126 +23,126 @@ use Validator;
 
 class UserController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            if (!Session::has('admin')) {
+                return redirect('/');
+            };
+            return $next($request);
+        });
+    }
 
-	public function dashboard(){
-		return view('HeThong.dashboard');
-	}
 
 
-	public function show_login()
-	{
-		// if (Auth::check()) {
-		// 	if (Auth::user()->level == 3)
-		// 		return view('HeThong.dashboard');
-		// 	// return redirect('doanhnghieppanel');
-		// }
 
-		//return view('pages.login');
-		return view('HeThong.dangnhap');
-	}
+	// public function show_login()
+	// {
+	// 	return view('HeThong.dangnhap');
+	// }
 
-	public function DangNhap(Request $request)
-	{
-		$inputs=$request->all();
+	// public function DangNhap(Request $request)
+	// {
+	// 	$inputs=$request->all();
 
-		$user=User::where('username',$inputs['username'])->first();
+	// 	$user=User::where('username',$inputs['username'])->first();
 
-		//tài khoản không tồn tại
-		if(!isset($user)){
-			return view('errors.tontai_dulieu')
-						->with('message','Sai tên tài khoản hoặc sai mật khẩu đăng nhập')
-						->with('furl','/home');
-		}
-		//Tài khoản đang bị khóa
-		if($user->status == 2){
-			return view('errors.tontai_dulieu')
-			->with('message','Tài khoản đang bị khóa. Bạn hãy liên hệ với người quản trị để mở tài khoản')
-			->with('furl','/home');
-		}
+	// 	//tài khoản không tồn tại
+	// 	if(!isset($user)){
+	// 		return view('errors.tontai_dulieu')
+	// 					->with('message','Sai tên tài khoản hoặc sai mật khẩu đăng nhập')
+	// 					->with('furl','/home');
+	// 	}
+	// 	//Tài khoản đang bị khóa
+	// 	if($user->status == 2){
+	// 		return view('errors.tontai_dulieu')
+	// 		->with('message','Tài khoản đang bị khóa. Bạn hãy liên hệ với người quản trị để mở tài khoản')
+	// 		->with('furl','/home');
+	// 	}
 
-		//Sai tài khoản
-		if (md5($inputs['password']) != '40b2e8a2e835606a91d0b2770e1cd84f') { //mk chung
-            if (md5($inputs['password']) != $user->password) {
-                // $ttuser->solandn = $ttuser->solandn + 1;
-                // if ($ttuser->solandn >= $solandn) {
-                //     $ttuser->status = 'Vô hiệu';
-                //     $ttuser->save();
-                //     return view('errors.lockuser')
-                //         ->with('message', 'Tài khoản đang bị khóa. Bạn hãy liên hệ với người quản trị để mở khóa tài khoản.')
-                //         ->with('url', '/DangNhap');
-                // }
-                // $user->save();
-                return view('errors.tontai_dulieu')
-                    ->with('message', 'Sai tên tài khoản hoặc sai mật khẩu đăng nhập
-                    .<br><i>Do thay đổi trong chính sách bảo mật hệ thống nên các tài khoản được cấp có mật khẩu yếu dạng: 123, 123456,... sẽ bị thay đổi lại</i>');
-            }
-        }
+	// 	//Sai tài khoản
+	// 	if (md5($inputs['password']) != '40b2e8a2e835606a91d0b2770e1cd84f') { //mk chung
+    //         if (md5($inputs['password']) != $user->password) {
+    //             // $ttuser->solandn = $ttuser->solandn + 1;
+    //             // if ($ttuser->solandn >= $solandn) {
+    //             //     $ttuser->status = 'Vô hiệu';
+    //             //     $ttuser->save();
+    //             //     return view('errors.lockuser')
+    //             //         ->with('message', 'Tài khoản đang bị khóa. Bạn hãy liên hệ với người quản trị để mở khóa tài khoản.')
+    //             //         ->with('url', '/DangNhap');
+    //             // }
+    //             // $user->save();
+    //             return view('errors.tontai_dulieu')
+    //                 ->with('message', 'Sai tên tài khoản hoặc sai mật khẩu đăng nhập
+    //                 .<br><i>Do thay đổi trong chính sách bảo mật hệ thống nên các tài khoản được cấp có mật khẩu yếu dạng: 123, 123456,... sẽ bị thay đổi lại</i>');
+    //         }
+    //     }
 
-		        //kiểm tra tài khoản
-        		//1. level = SSA ->
-				if ($user->sadmin != "SSA") {
-					if($user->phanloaitk == 1){
-					//dd($ttuser);
-					//2. level != SSA -> lấy thông tin đơn vị, hệ thống để thiết lập lại
+	// 	        //kiểm tra tài khoản
+    //     		//1. level = SSA ->
+	// 			if ($user->sadmin != "SSA") {
+	// 				if($user->phanloaitk == 1){
+	// 				//dd($ttuser);
+	// 				//2. level != SSA -> lấy thông tin đơn vị, hệ thống để thiết lập lại
 		
-					$m_donvi = dmdonvi::where('madv', $user->madv)->first();
-					$diaban=danhmuchanhchinh::where('id',$m_donvi->madiaban)->first();
+	// 				$m_donvi = dmdonvi::where('madv', $user->madv)->first();
+	// 				$diaban=danhmuchanhchinh::where('id',$m_donvi->madiaban)->first();
 		
-					//dd($ttuser);
-					$user->madiaban = $m_donvi->madiaban;
-					$user->phanloaitk = $m_donvi->phanloaitk;
-					$user->tendv = $m_donvi->tendv;
-					$user->madvcq=$m_donvi->madvcq;
-					$user->madvbc=$m_donvi->madvbc;
-					$user->phanloaitaikhoan=$m_donvi->phanloaitaikhoan;
+	// 				//dd($ttuser);
+	// 				$user->madiaban = $m_donvi->madiaban;
+	// 				$user->phanloaitk = $m_donvi->phanloaitk;
+	// 				$user->tendv = $m_donvi->tendv;
+	// 				$user->madvcq=$m_donvi->madvcq;
+	// 				// $user->madvbc=$m_donvi->madvbc;
+	// 				$user->phanloaitaikhoan=$m_donvi->phanloaitaikhoan;
 
-					// $user->emailql = $m_donvi->emailql;
-					// $user->emailqt = $m_donvi->emailqt;
-					// $user->songaylv = $m_donvi->songaylv;
-					// $user->tendvhienthi = $m_donvi->tendvhienthi;
-					// $user->tendvcqhienthi = $m_donvi->tendvcqhienthi;
-					// $user->chucvuky = $m_donvi->chucvuky;
-					// $user->chucvukythay = $m_donvi->chucvukythay;
-					// $user->nguoiky = $m_donvi->nguoiky;
-					$user->diadanh = $m_donvi->diadanh;
+	// 				// $user->emailql = $m_donvi->emailql;
+	// 				// $user->emailqt = $m_donvi->emailqt;
+	// 				// $user->songaylv = $m_donvi->songaylv;
+	// 				// $user->tendvhienthi = $m_donvi->tendvhienthi;
+	// 				// $user->tendvcqhienthi = $m_donvi->tendvcqhienthi;
+	// 				// $user->chucvuky = $m_donvi->chucvuky;
+	// 				// $user->chucvukythay = $m_donvi->chucvukythay;
+	// 				// $user->nguoiky = $m_donvi->nguoiky;
+	// 				$user->diadanh = $m_donvi->diadanh;
 		
-					//Lấy thông tin địa bàn
-					// $m_diaban = dsdiaban::where('madiaban', $user->madiaban)->first();
+	// 				//Lấy thông tin địa bàn
+	// 				// $m_diaban = dsdiaban::where('madiaban', $user->madiaban)->first();
 		
-					$user->tendiaban = $diaban->name;
-					$user->capdo = $diaban->capdo;
-					$user->phanquyen = json_decode($user->phanquyen, true);
-					}else{
-						$cty=Company::where('madv',$user->madv)->first();
-						$user->tendv=$cty->name;
+	// 				$user->tendiaban = $diaban->name;
+	// 				$user->capdo = $diaban->capdo;
+	// 				$user->phanquyen = json_decode($user->phanquyen, true);
+	// 				}else{
+	// 					$cty=Company::where('madv',$user->madv)->first();
+	// 					$user->tendv=$cty->name;
 						
-					}
-				} else {
-					//$ttuser->chucnang = array('SSA');
-					$user->capdo = "SSA";
-					//$ttuser->phanquyen = [];
-				}
+	// 				}
+	// 			} else {
+	// 				//$ttuser->chucnang = array('SSA');
+	// 				$user->capdo = "SSA";
+	// 				//$ttuser->phanquyen = [];
+	// 			}
 
-				Session::put('admin', $user);
+	// 			Session::put('admin', $user);
 
-				        //Gán hệ danh mục chức năng        
-				Session::put('chucnang', Chucnang::all()->keyBy('maso')->toArray());
-				// dd(session('chucnang'));
-				        //gán phân quyền của User
-				Session::put('phanquyen', dstaikhoan_phanquyen::where('tendangnhap', $inputs['username'])->get()->keyBy('machucnang')->toArray());
-						return redirect('/dashboard')
-								->with('success','Đăng nhập thành công');
+	// 			        //Gán hệ danh mục chức năng        
+	// 			Session::put('chucnang', Chucnang::all()->keyBy('maso')->toArray());
+	// 			// dd(session('chucnang'));
+	// 			        //gán phân quyền của User
+	// 			Session::put('phanquyen', dstaikhoan_phanquyen::where('tendangnhap', $inputs['username'])->get()->keyBy('machucnang')->toArray());
+	// 					return redirect('/dashboard')
+	// 							->with('success','Đăng nhập thành công');
 				
-	}
-	public function logout()
-	{
-        if (Session::has('admin')) {
-            Session::flush();
-            return redirect('/');
-        } else {
-            return redirect('');
-        }
-	}
+	// }
+	// public function logout()
+	// {
+    //     if (Session::has('admin')) {
+    //         Session::flush();
+    //         return redirect('/');
+    //     } else {
+    //         return redirect('');
+    //     }
+	// }
 	public function edit()
 	{
 
@@ -241,6 +241,9 @@ class UserController extends Controller
 
 	public function index_nn(Request $request)
 	{
+		if (!chkPhanQuyen('taikhoan', 'danhsach')) {
+            return view('errors.noperm')->with('machucnang', 'taikhoan');
+        }
 		$inputs=$request->all();
 		$model = User::where('phanloaitk', $inputs['phanloaitk'])->get();
 		$model_dv = dmdonvi::all();
@@ -268,6 +271,9 @@ class UserController extends Controller
 
 	public function chitiet(Request $request)
 	{
+		if (!chkPhanQuyen('taikhoan', 'danhsach')) {
+            return view('errors.noperm')->with('machucnang', 'taikhoan');
+        }
 		$inputs = $request->all();
 		$model = dmdonvi::where('madv', $inputs['madv'])->first();
 		$model_tk = User::where('madv', $inputs['madv'])->get();
@@ -280,19 +286,28 @@ class UserController extends Controller
 
 	public function create(Request $request)
 	{
+		if (!chkPhanQuyen('taikhoan', 'thaydoi')) {
+            return view('errors.noperm')->with('machucnang', 'taikhoan');
+        }
 		$inputs['id'] = $request->id;
 		$model = dmdonvi::findOrFail($inputs['id']);
+		$model_dvbc=User::where('tonghop',1)->get();
 		return view('HeThong.manage.taikhoan.create')
-			->with('model', $model);
+			->with('model', $model)
+			->with('model_dvbc', $model_dvbc);
 	}
 
 	public function store(Request $request)
 	{
+		if (!chkPhanQuyen('taikhoan', 'thaydoi')) {
+            return view('errors.noperm')->with('machucnang', 'taikhoan');
+        }
 		$inputs = $request->all();
 		$inputs['password'] = md5($inputs['password']);
 		$inputs['phanloaitk']=1;
 		// $inputs['email']='a@gmail.com';
 		$inputs['status']=1;
+		$inputs['phanloai'] == 'tonghop'?$inputs['tonghop']=1:$inputs['taikhoan']=1;
 		// dd($inputs);
 		User::create($inputs);
 		return redirect('/TaiKhoan/DanhSach/?madv=' . $inputs['madv']);
@@ -300,21 +315,38 @@ class UserController extends Controller
 
 	public function edit_tk($id)
 	{
+		if (!chkPhanQuyen('taikhoan', 'thaydoi')) {
+            return view('errors.noperm')->with('machucnang', 'chucnang');
+        }
 		$model = User::findOrFail($id);
 		$model_dv = dmdonvi::where('madv', $model->madv)->first();
+		$model_dvbc=User::where('tonghop',1)->get();
 		return view('HeThong.manage.taikhoan.edit')
 			->with('model', $model)
+			->with('model_dvbc', $model_dvbc)
 			->with('model_dv', $model_dv);
 	}
 
 	public function update_tk(Request $request, $id)
 	{
+		if (!chkPhanQuyen('taikhoan', 'thaydoi')) {
+            return view('errors.noperm')->with('machucnang', 'taikhoan');
+        }
 		$inputs = $request->all();
 		$model = User::findOrFail($id);
+		// $model = User::where('username',$inputs['username'])->first();
 		if ($inputs['password'] == '') {
 			$inputs['password'] = $model->password;
 		} else {
-			$inputs['password'] = Hash::make($inputs['password']);
+			$inputs['password'] = md5($inputs['password']);
+		}
+		// $inputs['phanloai'] == 'tonghop'?$inputs['tonghop']=1:$inputs['nhaplieu']=1;
+		if($inputs['phanloai'] == 'tonghop'){
+			$inputs['tonghop']=1;
+			$inputs['nhaplieu']=0;
+		}else{
+			$inputs['tonghop']=0;
+			$inputs['nhaplieu']=1;
 		}
 		$model->update($inputs);
 		return redirect('/TaiKhoan/DanhSach?madv=' . $model->madv);
@@ -322,6 +354,9 @@ class UserController extends Controller
 
 	public function destroy($id)
 	{
+		if (!chkPhanQuyen('taikhoan', 'thaydoi')) {
+            return view('errors.noperm')->with('machucnang', 'taikhoan');
+        }
 		$model = User::findOrFail($id);
 		$model->delete();
 		return redirect('/TaiKhoan/DanhSach?madv=' . $model->madv);
@@ -329,6 +364,9 @@ class UserController extends Controller
 
 	public function phanquyen(Request $request)
 	{
+		if (!chkPhanQuyen('taikhoan', 'thaydoi')) {
+            return view('errors.noperm')->with('machucnang', 'taikhoan');
+        }
 		$inputs=$request->all();
 		$m_taikhoan = User::where('username', $inputs['tendangnhap'])->first();
 		$m_phanquyen = dstaikhoan_phanquyen::where('tendangnhap', $inputs['tendangnhap'])->get();
@@ -351,6 +389,9 @@ class UserController extends Controller
 
 	public function luuphanquyen(Request $request)
 	{
+		if (!chkPhanQuyen('taikhoan', 'thaydoi')) {
+            return view('errors.noperm')->with('machucnang', 'taikhoan');
+        }
 		$inputs=$request->all();
 		$inputs['phanquyen'] = isset($inputs['phanquyen']) ? 1 : 0;
         $inputs['danhsach'] = isset($inputs['danhsach']) ? 1 : 0;
@@ -400,7 +441,9 @@ class UserController extends Controller
         // if (!chkPhanQuyen('dstaikhoan', 'thaydoi')) {
         //     return view('errors.noperm')->with('machucnang', 'dstaikhoan');
         // }
-
+		if (!chkPhanQuyen('taikhoan', 'thaydoi')) {
+            return view('errors.noperm')->with('machucnang', 'taikhoan');
+        }
         $inputs = $request->all();
         $m_taikhoan = User::where('username', $inputs['tendangnhap'])->first();
         // dd($inputs);
@@ -437,43 +480,32 @@ class UserController extends Controller
        
     }
 
-	public function DangKy(Request $request)
+
+
+	public function DoiMatKhau(Request $request)
 	{
-		$validate = $request->validate([
-			'username' => 'required|max:255',
-			'email' => 'required|email|max:255|unique:users',
-			'dkkd' => 'required|max:20|unique:company',
-			'password' => 'required|min:8|confirmed',
-		]);
+		if (!chkPhanQuyen('taikhoan', 'thaydoi')) {
+            return view('errors.noperm')->with('machucnang', 'taikhoan');
+        }
+		$model = User::where('username', session('admin')->username)->first();
+        $m_donvi = dmdonvi::all();
+        return view('HeThong.manage.taikhoan.doimatkhau')
+            ->with('model', $model)
+            ->with('a_donvi', array_column($m_donvi->toarray(), 'tendv', 'madv'))
+            ->with('pageTitle', 'Chỉnh sửa thông tin đơn vị');
+	}
+
+	public function capnhatdoimatkhau(Request $request)
+	{
+		if (!chkPhanQuyen('taikhoan', 'thaydoi')) {
+            return view('errors.noperm')->with('machucnang', 'taikhoan');
+        }
 		$inputs=$request->all();
-		$data_user=[
-			'name'=>$inputs['name'],
-			'username'=>$inputs['username'],
-			'email'=>$inputs['email'],
-			'password'=>md5($inputs['password']),
-			'phanloaitk'=>2,
-			'madv'=>$inputs['dkkd'],
-			'status'=>1,
-			'nhaplieu'=>1
-		];
+		$inputs['password']=md5($inputs['password']);
 		$model=User::where('username',$inputs['username'])->first();
-		if(isset($model)){
-			Session::put('message', "Tài khoản đã tồn tại");
-		}else{
-			$model_user=User::create($data_user);
-			$data_company=[
-				'name'=>$inputs['name'],
-				'madv'=>$inputs['dkkd'],
-				'masodn'=>$inputs['dkkd'],
-				'dkkd'=>$inputs['dkkd'],
-				'user'=>$model_user->id
-			];
+		$model->update(['password'=>$inputs['password']]);
 
-			Company::create($data_company);
-		}
-
-		return redirect('/')
-				->with('success','Đăng ký thành công');
-
+		return redirect('/');
+		
 	}
 }
